requirements.txt
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
playwright==1.40.0
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.11
requests==2.31.0